# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's Moon Group Transparency",
        "description": "API's for transparency of Moon Group",
        "version": "1.0.0"
    },
    "paths": {
        "/moontransparency/public/api/activity": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Obtener información de Activitys con filtros y ordenamiento",
                "operationId": "a58d42bb4654f8993c6fc2d25785263c",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre de Activity",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filtrar por fecha de inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filtrar por fecha de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Filtrar por ID del proyecto",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "objective",
                        "in": "query",
                        "description": "Filtrar por objetivo de la actividad",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "total_amount",
                        "in": "query",
                        "description": "Filtrar por monto total de la actividad",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "collected_amount",
                        "in": "query",
                        "description": "Filtrar por monto recolectado",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado de la actividad",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Activitys",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activity"
                ],
                "summary": "Crear Activity",
                "operationId": "43230c5172e418969d6e915bbf2fc69f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "La validación falló."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moontransparency/public/api/activity/{id}": {
            "get": {
                "tags": [
                    "Activity"
                ],
                "summary": "Obtener detalles de un Activity por ID",
                "operationId": "d20fe9bf1ffccc58590e0f95798f466c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Activity",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donación no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Activity"
                ],
                "summary": "Actualizar un Activity",
                "operationId": "ccc0c4e2036d1c05745fa418e86d0c46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donación actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donación no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activity"
                ],
                "summary": "Eliminar un Activity por ID",
                "operationId": "9622fe8f59f33594d7a19d54261d293c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Donación eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Donación no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ActivityRequest": {
                "required": [
                    "project_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Proyecto de Energía Renovable"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "project_id": {
                        "type": "integer",
                        "example": "101"
                    },
                    "objective": {
                        "type": "string",
                        "example": "Implementar fuentes de energía renovable."
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": "500000"
                    },
                    "collected_amount": {
                        "type": "number",
                        "format": "float",
                        "example": "200000"
                    }
                },
                "type": "object"
            },
            "Activity": {
                "title": "Activity",
                "description": "Modelo de Activity",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Proyecto de Energía Renovable"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "project_id": {
                        "type": "integer",
                        "example": "101"
                    },
                    "objective": {
                        "type": "string",
                        "example": "Implementar fuentes de energía renovable."
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": "500000"
                    },
                    "collected_amount": {
                        "type": "number",
                        "format": "float",
                        "example": "200000"
                    },
                    "status": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-26T21:44:24"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}